IOS/ES: Move GetStoredContentsFromTMD to NandUtils
